Update App-Auto-Patch-via-Dialog.zsh with Deferral Function
Added a "checkDeferral" function that allows you to prompt a user to install pending updates or defer for a set number of times defined by the maxDeferrals variable. Setting this variable to 
Disabled will prevent this dialog from appearing. 
There is a deferralTimer variable that will present a countdown for how long the user has to select an option, otherwise installation will continue