Update App-Auto-Patch-via-Dialog.zsh
Added a script variable and code for ignoring applications installed in the /Users/* hierarchy.

When App-Auto-Update finds an app that is installed in /Users/* and asks Installomator to update it, the machine generally ends up with two copies of the application. The current version installed in the /Applications folder and the downlevel version still install in the original /Users/* location. This gives the ability to simply ignore those apps.