Added code for switching the Installomator from Release to Beta
This will allow the administrator to choose either the latest release version of Installomator or the current beta (main branch).

Any approved PRs for new labels, or bug fixes to the Installomator code would be added to the main branch sooner than waiting for another release. In some cases, people may prefer to use the main branch over the latest release of Installomator.