Remove hard coded "/Library/Managment/" references and unecessary chmod/chown
Removed:
2201 chown root: wheel "/Library/Management"
2202 chmod 777 "/Library/Management"
Justification:
If /Library/Managment or any other directory is created by the script it will be "root: wheel" 755 by default.  If the folder exists with custom permissions, those shouldn't be changed. All other permissions for the "${appAutoPatchFolder}" directory and below are set elsewhere in the script.

Line 2350 and 2358, updated hard coded reference to "/Library/Managment/AppAutoPatch/Installomator/Installomater.sh" to "${installomatorScript}"
Justification: 
Script was failing to find and version "Installomator.sh" if a custom ${appAutoPatchFolder} was set.